+++ dof_discord_bot/src/constants.py
@@ -21,15 +21,15 @@ COMMAND_PREFIX = "!"
 # Declare the maximum number of the lines for the !help command
 MAX_HELP_LINES = 8
 
-# Declare the amount of seconds to wait until help session is deleted
-HELP_SESSION_TIMEOUT = 60
-
 # Declare some emojis
 DELETE_EMOJI = "\U0001f5d1"
 FIRST_PAGE_EMOJI = "\u23EE"
 PREVIOUS_PAGE_EMOJI = "\u2B05"
 NEXT_PAGE_EMOJI = "\u27A1"
 LAST_PAGE_EMOJI = "\u23ED"
+
+# Declare some icons
+DEFAULT_SESSION_ICON = "https://cdn.discordapp.com/emojis/512367613339369475.png"
 
 with open(_os.path.join(RES_DIR, "token.txt")) as f:
     TOKEN = f.read().strip()
