+++ dof_discord_bot/src/utils.py
@@ -2,8 +2,15 @@
 An unorganised collection of non-static constructs, such as classes or utility functions.
 """
 import discord as _discord
-from discord.ext import commands as _commands
+import abc as _abc
+import asyncio as _asyncio
+import contextlib as _contextlib
+from .logger import Log as _Log
 from . import strings as _strings
+from .constants import DEFAULT_SESSION_ICON as _DEFAULT_SESSION_ICON, LAST_PAGE_EMOJI as _LAST_PAGE_EMOJI, \
+    FIRST_PAGE_EMOJI as _FIRST_PAGE_EMOJI, NEXT_PAGE_EMOJI as _NEXT_PAGE_EMOJI, DELETE_EMOJI as _DELETE_EMOJI, \
+    PREVIOUS_PAGE_EMOJI as _PREVIOUS_PAGE_EMOJI
+from discord.ext import commands as _commands
 
 
 class MemberApplication:
@@ -109,7 +116,7 @@ class Page:
             self._lines.append(line)
 
     @property
-    def lines(self):
+    def lines(self) -> list:
         """
         Getter for the page lines.
         """
@@ -162,3 +169,263 @@ class LinePaginator(_commands.Paginator):
         """
         self.lines_count = 0
         super().close_page()
+
+
+class Session:
+    """
+    Interactive session used to format and display multi-paged text.
+
+    Each session is assigned to the user, and lasts a set amount of time (as specified by timeout), unless refreshed.
+
+    The pages can be browsed using the associated icons on the bottom of the message, or the message (and the session)
+    can be terminated early when used with the wastebasket icon.
+
+    When inheriting from this class, you must implement an asynchronous `build_pages` functions, and set the session's
+    pages in there.
+    """
+
+    def __init__(self, ctx: _commands.Context, title: str, icon: str = _DEFAULT_SESSION_ICON, timeout: int = 60):
+        """
+        Constructor is directly called by the `start` method, and always takes 3 arguments - context, title, and
+        optional icon.
+
+        If you want to pass additional information to the session (for example to build the pages using that
+        information), add it to passed context and override init to remember the value.
+
+        Title must be set and will be displayed with the message. Icon is optional and defaults to a question mark.
+
+        Timeout defines after how long of no interaction should the session be ended.
+        """
+        self.bot = ctx.bot
+        self.author = ctx.author
+        self.destination = ctx.channel
+        self.title = title
+        self.icon = icon
+        self.pages = list()
+        self.session_timeout = timeout
+        self.current_page = 0
+        self.message = None
+        self.timeout_task = None
+
+        # Declare a mapping of emoji to reaction functions
+        self.reactions = {
+            _FIRST_PAGE_EMOJI: self.do_first_page,
+            _PREVIOUS_PAGE_EMOJI: self.do_previous_page,
+            _NEXT_PAGE_EMOJI: self.do_next_page,
+            _LAST_PAGE_EMOJI: self.do_last_page,
+            _DELETE_EMOJI: self.do_delete
+        }
+
+    @classmethod
+    async def start(cls, ctx: _commands.Context, title: str, icon: str = _DEFAULT_SESSION_ICON) -> "Session":
+        """
+        Create and begin a session based on the given context.
+        """
+        _Log.info(f"Starting a session, initiated by {ctx.author}")
+
+        session = cls(ctx, title, icon)
+        await session.prepare()
+        return session
+
+    async def stop(self):
+        """
+        Stops the session, removes event listeners and attempts to delete the session message.
+        """
+        _Log.info(f"Stopping the session started by {self.author}")
+
+        self.bot.remove_listener(self.on_reaction_add)
+        self.bot.remove_listener(self.on_message_delete)
+
+        # Ignore if permission issue, or the message doesn't exist
+        with _contextlib.suppress(_discord.HTTPException, AttributeError):
+            await self.message.delete()
+
+    async def prepare(self):
+        """
+        Sets up the session pages, events, message, and reactions.
+        """
+        _Log.debug(f"Preparing the session for {self.author}")
+
+        # Create paginated content
+        await self.build_pages()
+
+        # Setup the listeners to allow page browsing
+        self.bot.add_listener(self.on_reaction_add)
+        self.bot.add_listener(self.on_message_delete)
+
+        # Display the first page
+        await self.update_page()
+
+        # Initial timeout reset to set the timer
+        self.reset_timeout()
+
+        # Add the reactions once the message is visible
+        self.add_reactions()
+
+    @_abc.abstractmethod
+    async def build_pages(self):
+        """
+        Method to be overridden - each session needs to know how to create the pages.
+
+        You MUST set self.pages in this method to a list of paginator's pages:
+
+            paginator = LinePaginator()
+            # Code which adds pages to paginator
+
+            self.pages = paginator.pages
+        """
+        pass
+
+    @property
+    def is_first_page(self) -> bool:
+        """
+        Check if session is currently showing the first page.
+        """
+        return self.current_page == 0
+
+    @property
+    def is_last_page(self) -> bool:
+        """
+        Check if the session is currently showing the last page.
+        """
+        return self.current_page == (len(self.pages) - 1)
+
+    async def do_first_page(self):
+        """
+        Event that is called when the user requests the first page.
+        """
+        _Log.debug(f"Getting first page for {self.author}")
+
+        if not self.is_first_page:
+            await self.update_page(0)
+
+    async def do_previous_page(self):
+        """
+        Event that is called when the user requests the previous page.
+        """
+        _Log.debug(f"Getting previous page for {self.author}")
+
+        if not self.is_first_page:
+            await self.update_page(self.current_page - 1)
+
+    async def do_next_page(self):
+        """
+        Event that is called when the user requests the next page.
+        """
+        _Log.debug(f"Getting next page for {self.author}")
+
+        if not self.is_last_page:
+            await self.update_page(self.current_page + 1)
+
+    async def do_last_page(self):
+        """
+        Event that is called when the user requests the last page.
+        """
+        _Log.debug(f"Getting last page for {self.author}")
+
+        if not self.is_last_page:
+            await self.update_page(len(self.pages) - 1)
+
+    async def do_delete(self):
+        """
+        Event that is called when the user requests to stop the help session.
+        """
+        _Log.debug(f"Deleting the message for {self.author}")
+
+        await self.message.delete()
+
+    async def timeout(self):
+        """
+        Waits for a set number of seconds, then stops the help session.
+        """
+        await _asyncio.sleep(self.session_timeout)
+        await self.stop()
+
+    def reset_timeout(self):
+        """
+        Cancels the original timeout task and sets it up again from the start.
+
+        Used mainly to keep the session after users interact with it.
+        """
+        _Log.debug(f"A user action by {self.author} forced the timeout reset")
+
+        # Cancel the original task if it exists
+        if self.timeout_task:
+            if not self.timeout_task.cancelled():
+                self.timeout_task.cancel()
+
+        # Recreate the timeout task
+        self.timeout_task = self.bot.loop.create_task(self.timeout())
+
+    def add_reactions(self):
+        """
+        Adds the relevant reactions to the session message based on if pagination is required.
+        """
+        _Log.debug(f"Adding reactions for {self.author}'s session")
+
+        if len(self.pages) > 1:
+            for reaction in self.reactions:
+                self.bot.loop.create_task(self.message.add_reaction(reaction))
+        else:
+            self.bot.loop.create_task(self.message.add_reaction(_DELETE_EMOJI))
+
+    async def update_page(self, page_number: int = 0):
+        """
+        Displays the initial page, or changes the existing one to the given page number.
+        """
+        self.current_page = page_number
+        embed_page = self.embed_page(page_number)
+
+        if not self.message:
+            self.message = await self.destination.send(embed=embed_page)
+        else:
+            await self.message.edit(embed=embed_page)
+
+    def embed_page(self, page_number: int = 0) -> _discord.Embed:
+        """
+        Returns an Embed with the requested page formatted within.
+        """
+        embed = _discord.Embed()
+
+        embed.set_author(name=self.title, icon_url=self.icon)
+        embed.description = self.pages[page_number]
+
+        # Add page counter to footer if paginating
+        page_count = len(self.pages)
+        if page_count > 1:
+            embed.set_footer(text=f'Page {self.current_page + 1} / {page_count}')
+
+        return embed
+
+    async def on_reaction_add(self, reaction: _discord.Reaction, user: _discord.User):
+        """
+        Event handler for when reactions are added on the session message.
+        """
+        _Log.debug(f"Reaction added by {user.display_name}")
+
+        # Ensure it was the relevant session message
+        if reaction.message.id != self.message.id:
+            return
+
+        # Ensure it was the session author who reacted
+        if user.id != self.author.id:
+            return
+
+        # Only handle valid action emoji-s
+        if str(reaction.emoji) in self.reactions:
+            self.reset_timeout()
+            await self.reactions[str(reaction.emoji)]()
+        else:
+            return
+
+        # Remove the added reaction to prep for re-use
+        with _contextlib.suppress(_discord.HTTPException):
+            _Log.debug(f"Reaction by {user.display_name} handled by the session")
+            await self.message.remove_reaction(reaction, user)
+
+    async def on_message_delete(self, message: _discord.Message):
+        """
+        Closes the help session when the session message is deleted.
+        """
+        if message.id == self.message.id:
+            await self.stop()
